# apps/til-backend/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Today I Learned Backend API
  version: 0.1.0
paths:
  /user-info:
    get:
      summary: User Info
      operationId: user_info_user_info_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /submissions:
    get:
      summary: List Submissions
      operationId: list_submissions_submissions_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Submission
      operationId: create_submission_submissions_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /submissions/{submission_id}:
    delete:
      summary: Delete Submission
      operationId: delete_submission_submissions__submission_id__delete
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /health:
    get:
      summary: Health
      operationId: health_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError